requirements.txt
flask==2.3.3
streamlit==1.28.1
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
requests==2.31.0
beautifulsoup4==4.12.2
yfinance==0.2.18
gtts==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2 